heartbeat/mysql - MAJOR CHANGE: always call check_slave for slave role if we are in a master/slave mode
If we do NOT call check_slave (in the old version required very specific set of configuration options) we can end up in very, very, very bad scenario where broken or outdated slave gets promoted as master.

This is very dangerous and very hard to resolve at later stage as if you end up in situation like this you will have to manually merge databases accross freshly-failed-master and outdated-slave-and-later-promoted-to-be-master

This also lead to code simplification

In previous commits check_slave has been extended to detect replication errors in Last_SQL_Error/Last_SQL_Errno, Last_IO_Error/Last_IO_Errno, Last_Error/Last_Errno which indicate that the slave is NOT operational and up2date thus useless or dangerous to be used as most probbably data loss will ocur if we use it as master.

Other minor changes in this commit:

	- separated master_pref scoring in check_slave to be outside if statements
	- mysql_monitor should ALWAYS call check_slave if running in MS mode regardless of the fact if $OCF_RESKEY_test_table is set or not
	- renamed secs_behind to Seconds_Behind_Master so we can clearly see what is held in this var and from where it comes
	- call $CRM_MASTER -D and set_reader_attr 0 in quit_on_replication_error only if we are running in ms mode
	- simplified quit_on_replication_error
	- as $tmpfile from show slave status is removed then do not show it in logs
	- merged $Slave_SQL_Running and $Slave_IO_Running checks in single if statement where in both cases we attempt to start slave